Refactor User tests
